Configuring DbContext - Implement MediatR

+++ Dominio/Price.cs
@@ -2,10 +2,11 @@ namespace Dominio
 {
     public class Price
     {
+        
         public int PrecioId { get; set; }
         public decimal ActualPrice { get; set; }
         public decimal Promotion { get; set; }
-        public decimal CourseId { get; set; }
+        public int CourseId { get; set; }
         public Course Course { get; set; }
     }
 }
